xtee6/common-util/src/main/java/ee/ria/xroad/common/certificateprofile/impl/EjbcaCertificateProfileInfoProvider.java
package ee.ria.xroad.common.certificateprofile.impl;

import ee.ria.xroad.common.certificateprofile.AuthCertificateProfileInfo;
import ee.ria.xroad.common.certificateprofile.CertificateProfileInfoProvider;
import ee.ria.xroad.common.certificateprofile.SignCertificateProfileInfo;

/**
 * Default implementation of CertificateProfileInfoProvider.
 */
public class EjbcaCertificateProfileInfoProvider
        implements CertificateProfileInfoProvider {

    @Override
    public AuthCertificateProfileInfo getAuthCertProfile(
            AuthCertificateProfileInfo.Parameters params) {
        return new EjbcaAuthCertificateProfileInfo(params);
    }

    @Override
    public SignCertificateProfileInfo getSignCertProfile(
            SignCertificateProfileInfo.Parameters params) {
        return new EjbcaSignCertificateProfileInfo(params);
    }

}
